[refactor]: RoleEnum 의 user -> core 모듈로의 리펙토링
- 여러 엔티티의 컨트롤러에서 호출할 UserRoleCheckerImpl 의 RoleEnum 의존성을 위한 리펙토링
