+++ user/src/main/java/com/catpang/user/domain/User.java
@@ -1,6 +1,16 @@
 package com.catpang.user.domain;
 
-import jakarta.persistence.*;
+import com.catpang.core.domain.model.RoleEnum;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.DiscriminatorColumn;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,18 +25,18 @@ import lombok.NoArgsConstructor;
 @Table(name = "p_users")
 public class User {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id", nullable = false)
-    private Long id;
-    @Column(nullable = false)
-    private String name;
-    @Column(nullable = false)
-    private String nickname;
-    @Column(nullable = false)
-    private String password;
-    @Column(nullable = false)
-    private String email;
-    @Enumerated(EnumType.STRING)
-    private RoleEnum role;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "user_id", nullable = false)
+	private Long id;
+	@Column(nullable = false)
+	private String name;
+	@Column(nullable = false)
+	private String nickname;
+	@Column(nullable = false)
+	private String password;
+	@Column(nullable = false)
+	private String email;
+	@Enumerated(EnumType.STRING)
+	private RoleEnum role;
 }
